+++ packages/tendermint-rpc/types/index.d.ts
@@ -1,4 +1,5 @@
-export { v0_33 } from "./v0-33";
+export { Adaptor } from "./adaptor";
+export { adaptor33, adaptor34 } from "./adaptors";
 export { Client } from "./client";
 export {
   AbciInfoRequest,